+++ pronamic-pay-woocommerce.php
@@ -4,7 +4,7 @@
  * Plugin URI: https://www.pronamic.eu/plugins/pronamic-pay-woocommerce/
  * Description: Extend the Pronamic Pay plugin with WooCommerce support to receive payments through a variety of payment providers.
  *
- * Version: 4.8.0
+ * Version: 4.12.1
  * Requires at least: 5.9
  * Requires PHP: 7.4
  *
@@ -19,7 +19,7 @@
  * GitHub URI: https://github.com/pronamic/wp-pronamic-pay-woocommerce
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2023 Pronamic
+ * @copyright 2005-2025 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Extensions\WooCommerce
  */
@@ -37,7 +37,7 @@ require_once __DIR__ . '/vendor/autoload_packages.php';
  * Bootstrap.
  */
 add_action(
-	'plugins_loaded',
+	'init',
 	function () {
 		load_plugin_textdomain( 'pronamic-pay-woocommerce', false, dirname( plugin_basename( __FILE__ ) ) . '/languages' );
 	}
@@ -64,3 +64,14 @@ add_filter(
 		return $integrations;
 	}
 );
+
+if ( class_exists( \Pronamic\WordPress\Pay\Gateways\Mollie\Integration::class ) ) {
+	add_filter(
+		'pronamic_pay_gateways',
+		function ( $gateways ) {
+			$gateways[] = new \Pronamic\WordPress\Pay\Gateways\Mollie\Integration();
+
+			return $gateways;
+		}
+	);
+}