+++ src/WooCommerceSubscriptionsController.php
@@ -3,7 +3,7 @@
  * WooCommerce Subscriptions controller
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2023 Pronamic
+ * @copyright 2005-2025 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Extensions\WooCommerce
  */
@@ -91,7 +91,7 @@ class WooCommerceSubscriptionsController {
 
 		$subscription_edit_link = \sprintf(
 			/* translators: %s: order number */
-			\__( 'Subscription %s', 'pronamic_ideal' ),
+			\__( 'Subscription %s', 'pronamic-pay-woocommerce' ),
 			$source_id
 		);
 
@@ -107,7 +107,7 @@ class WooCommerceSubscriptionsController {
 						$edit_post_url,
 						\sprintf(
 							/* translators: %s: order number */
-							\__( 'Subscription %s', 'pronamic_ideal' ),
+							\__( 'Subscription %s', 'pronamic-pay-woocommerce' ),
 							$woocommerce_subscription->get_order_number()
 						),
 					);
@@ -116,7 +116,7 @@ class WooCommerceSubscriptionsController {
 		}
 
 		$text = [
-			\__( 'WooCommerce', 'pronamic_ideal' ),
+			\__( 'WooCommerce', 'pronamic-pay-woocommerce' ),
 			$subscription_edit_link,
 		];
 
@@ -130,7 +130,7 @@ class WooCommerceSubscriptionsController {
 	 * @return string
 	 */
 	public static function subscription_source_description( $description ) {
-		$description = __( 'WooCommerce Subscription', 'pronamic_ideal' );
+		$description = __( 'WooCommerce Subscription', 'pronamic-pay-woocommerce' );
 
 		return $description;
 	}
@@ -231,7 +231,7 @@ class WooCommerceSubscriptionsController {
 
 	/**
 	 * Maybe add a Pronamic Pay meta box the WooCommerce subscription.
-	 * 
+	 *
 	 * @link https://github.com/pronamic/wp-pronamic-pay-woocommerce/issues/41
 	 * @link https://developer.wordpress.org/reference/hooks/add_meta_boxes/
 	 * @param string           $post_type_or_screen_id Post type or screen ID.
@@ -251,7 +251,7 @@ class WooCommerceSubscriptionsController {
 
 		\add_meta_box(
 			'woocommerce-subscription-pronamic-pay',
-			\__( 'Pronamic Pay', 'pronamic_ideal' ),
+			\__( 'Pronamic Pay', 'pronamic-pay-woocommerce' ),
 			function () use ( $subscription ) {
 				include __DIR__ . '/../views/admin-meta-box-woocommerce-subscription.php';
 			},